very good improvement from cyanox

+++ src/app/ModuleBrowser.cpp
@@ -192,7 +192,7 @@ struct ModelBox : widget::OpaqueWidget {
 	}
 
 	void step() override {
-		if (previewFb && ++visibleFrames >= 60) {
+		if (!settings::keepPreviews && previewFb && ++visibleFrames >= 60) {
 			deletePreview();
 		}
 		OpaqueWidget::step();
